[fix]: RegionBottomSheet 위치 변경
