+++ feature/login/src/main/java/com/core/screen/InformationScreen.kt
@@ -12,40 +12,30 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.foundation.lazy.grid.items
 import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.ModalBottomSheet
 import androidx.compose.material3.Text
-import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalFocusManager
-import androidx.compose.ui.res.stringArrayResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.core.model.signup.SignType
 import com.youth.app.feature.login.R
 import com.youthtalk.component.button.CheckButton
-import com.youthtalk.component.button.SelectedButton
 import com.youthtalk.component.dropdown.RegionDropDown
 import com.youthtalk.component.inputfield.InputTextField
+import com.youthtalk.component.sheet.RegionBottomSheet
 import com.youthtalk.designsystem.YongProjectTheme
-import com.youthtalk.designsystem.gray10
 import com.youthtalk.designsystem.gray100
 import com.youthtalk.designsystem.gray80
-import kotlinx.coroutines.launch
 import timber.log.Timber
 
 @Composable
@@ -148,68 +138,6 @@ fun InformationScreen(loading: Boolean, onBack: () -> Unit, signUp: (String, Str
                 region = it
             }
         )
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun RegionBottomSheet(modifier: Modifier = Modifier, region: String, onDismiss: () -> Unit, onClick: (String) -> Unit) {
-    val regions = stringArrayResource(R.array.regions).toList()
-    var selectedRegion by remember {
-        mutableStateOf(region)
-    }
-    val scope = rememberCoroutineScope()
-    val sheetState = rememberModalBottomSheetState(
-        skipPartiallyExpanded = true
-    )
-    ModalBottomSheet(
-        modifier = modifier,
-        sheetState = sheetState,
-        onDismissRequest = onDismiss,
-        containerColor = gray10
-    ) {
-        Column(
-            modifier = Modifier
-                .padding(horizontal = 20.dp)
-        ) {
-            Text(
-                text = stringResource(R.string.bottom_sheet_title),
-                style = MaterialTheme.typography.bodyMedium.copy(
-                    color = gray100
-                )
-            )
-            LazyVerticalGrid(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(top = 20.dp),
-                columns = GridCells.Fixed(3),
-                horizontalArrangement = Arrangement.spacedBy(12.dp),
-                verticalArrangement = Arrangement.spacedBy(12.dp)
-            ) {
-                items(items = regions) {
-                    SelectedButton(
-                        text = it,
-                        isSelected = selectedRegion == it
-                    ) {
-                        selectedRegion = it
-                    }
-                }
-            }
-
-            CheckButton(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(vertical = 26.dp),
-                text = stringResource(R.string.bottom_sheet_button_text),
-                isCheck = selectedRegion.isNotEmpty()
-            ) {
-                scope.launch {
-                    sheetState.hide()
-                }
-                onClick(selectedRegion)
-                onDismiss()
-            }
-        }
     }
 }
